Java: GptUi: add temperature

+++ Java+/App+/GptUi/src/gptui/gpt/gcp/RequestBody.java
@@ -1,4 +1,6 @@
 package gptui.gpt.gcp;
 
-record RequestBody(RequestPrompt prompt, Double temperature, Integer candidateCount) {
+import java.math.BigDecimal;
+
+record RequestBody(RequestPrompt prompt, BigDecimal temperature, Integer candidateCount) {
 }
